Move Rails::RSpec as RSpec
