+++ lib/minipack/rails.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
-require 'logger'
-
 module Minipack
   module Rails
-    INSTALLER_WATCHED_PATHS = ['package.json', 'package-lock.json', 'yarn.lock'].freeze
-
     require 'minipack/configuration'
     require 'minipack/helper'
     require 'minipack/manifest_repository'
@@ -31,33 +27,6 @@ module Minipack
       # @deprecated
       def configuration=(c)
         Minipack.configuration = c
-      end
-
-      def install(logger: nil)
-        configuration.leaves.each do |c|
-          watched_paths = INSTALLER_WATCHED_PATHS.map { |f| File.expand_path(f, c.resolved_base_path) }
-          watcher = FileChangeWatcher.new(watched_paths, File.join(c.cache_path, "last-installation-digest-#{c.id}-#{::Rails.env}"))
-          CommandRunner.new(
-            {},
-            c.install_command,
-            chdir: c.resolved_base_path,
-            logger: logger,
-            watcher: watcher,
-          ).run!
-        end
-      end
-
-      def build(logger: nil)
-        configuration.leaves.each do |c|
-          watcher = FileChangeWatcher.new(c.resolved_watched_paths, File.join(c.cache_path, "last-build-digest-#{c.id}-#{::Rails.env}"))
-          CommandRunner.new(
-            {},
-            c.build_command,
-            chdir: c.resolved_base_path,
-            logger: logger,
-            watcher: watcher,
-          ).run!
-        end
       end
     end
   end
