+++ lib/minipack/rspec.rb
@@ -5,7 +5,7 @@ RSpec.configure do |c|
     logger = Logger.new(STDOUT).tap do |l|
       l.formatter = proc { |severity, datetime, progname, msg| "#{msg}\n" }
     end
-    Minipack::Rails.install(logger: logger)
-    Minipack::Rails.build(logger: logger)
+    Minipack.install(logger: logger)
+    Minipack.build(logger: logger)
   end
 end
